// src/modules/funds/funds.enums.ts
export enum FundMainCategory {
  EQUITY = 'Equity',
  DEBT = 'Debt',
  HYBRID = 'Hybrid',
  SOLUTION_ORIENTED = 'Solution Oriented',
  OTHER = 'Other',
}

export enum FundRiskLevel {
  LOW = 'Low',
  LOW_TO_MODERATE = 'Low to Moderate',
  MODERATE = 'Moderate',
  MODERATELY_HIGH = 'Moderately High',
  HIGH = 'High',
  VERY_HIGH = 'Very High',
}

export enum FundSubCategory {
  LARGE_CAP = 'Large Cap',
  MID_CAP = 'Mid Cap',
  SMALL_CAP = 'Small Cap',
  FLEXI_CAP = 'Flexi Cap',
  MULTI_CAP = 'Multi Cap',
  ELSS = 'ELSS',
  SECTORAL_THEMATIC = 'Sectoral/Thematic',
  LIQUID = 'Liquid',
  ULTRA_SHORT = 'Ultra Short',
  MONEY_MARKET = 'Money Market',
  ARBITRAGE = 'Arbitrage',
  DYNAMIC_ASSET_ALLOCATION = 'Dynamic Asset Allocation',
  INDEX_FUNDS = 'Index Funds',
  OTHER = 'Other',
}